@@ -1,9 +1,9 @@
 import 'dart:async';
 
-import 'package:HackRU/styles.dart';
-import 'package:HackRU/services/hackru_service.dart';
-import 'package:HackRU/models/models.dart';
-import 'package:HackRU/ui/pages/events/events_for_day.dart';
+import 'package:hackru/styles.dart';
+import 'package:hackru/services/hackru_service.dart';
+import 'package:hackru/models/models.dart';
+import 'package:hackru/ui/pages/events/events_for_day.dart';
 import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter/material.dart';
 
