@@ -1,8 +1,8 @@
-import 'package:HackRU/models/cred_manager.dart';
-import 'package:HackRU/styles.dart';
-import 'package:HackRU/ui/pages/dashboard/dashboard.dart';
-import 'package:HackRU/ui/pages/events/events.dart';
-import 'package:HackRU/ui/pages/login/login_page.dart';
+import 'package:hackru/models/cred_manager.dart';
+import 'package:hackru/styles.dart';
+import 'package:hackru/ui/pages/dashboard/dashboard.dart';
+import 'package:hackru/ui/pages/events/events.dart';
+import 'package:hackru/ui/pages/login/login_page.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 //import 'package:qr_flutter/qr_flutter.dart';
