@@ -1,9 +1,9 @@
-import 'package:HackRU/models/cred_manager.dart';
-import 'package:HackRU/services/hackru_service.dart';
-import 'package:HackRU/styles.dart';
-import 'package:HackRU/defaults.dart';
-import 'package:HackRU/ui/widgets/dialog/error_dialog.dart';
-import 'package:HackRU/ui/widgets/loading_indicator.dart';
+import 'package:hackru/models/cred_manager.dart';
+import 'package:hackru/services/hackru_service.dart';
+import 'package:hackru/styles.dart';
+import 'package:hackru/defaults.dart';
+import 'package:hackru/ui/widgets/dialog/error_dialog.dart';
+import 'package:hackru/ui/widgets/loading_indicator.dart';
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
 
