@@ -1,4 +1,4 @@
-import 'package:HackRU/styles.dart';
+import 'package:hackru/styles.dart';
 import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter/material.dart';
 
